core: Move emu_window and key_map into core
* Removes circular dependences (common should not depend on core)
