+++ src/citra/emu_window/emu_window_sdl2.cpp
@@ -9,10 +9,10 @@
 #include <SDL.h>
 #include <glad/glad.h>
 #include "citra/emu_window/emu_window_sdl2.h"
-#include "common/key_map.h"
 #include "common/logging/log.h"
 #include "common/scm_rev.h"
 #include "common/string_util.h"
+#include "core/frontend/key_map.h"
 #include "core/hle/service/hid/hid.h"
 #include "core/settings.h"
 #include "video_core/video_core.h"
