+++ src/core/CMakeLists.txt
@@ -29,6 +29,8 @@ set(SRCS
             file_sys/ivfc_archive.cpp
             file_sys/path_parser.cpp
             file_sys/savedata_archive.cpp
+            frontend/emu_window.cpp
+            frontend/key_map.cpp
             gdbstub/gdbstub.cpp
             hle/config_mem.cpp
             hle/applets/applet.cpp
@@ -198,6 +200,8 @@ set(HEADERS
             file_sys/ivfc_archive.h
             file_sys/path_parser.h
             file_sys/savedata_archive.h
+            frontend/emu_window.h
+            frontend/key_map.h
             gdbstub/gdbstub.h
             hle/config_mem.h
             hle/function_wrappers.h
