+++ src/core/frontend/emu_window.h
