+++ src/core/frontend/key_map.cpp
@@ -3,8 +3,8 @@
 // Refer to the license.txt file included.
 
 #include <map>
-#include "common/emu_window.h"
-#include "common/key_map.h"
+#include "core/frontend/emu_window.h"
+#include "core/frontend/key_map.h"
 
 namespace KeyMap {
